Move base classes to createConverter
